fix: preserve selection while scrolling
- track selection endpoints using logical scrollback rows
- allow wheel scrolling during active text selection
- auto-scroll near viewport edges while dragging
- copy selected text across visible and historical rows

+++ src/terminal/encoding.rs
@@ -5,7 +5,7 @@ use vt100::MouseProtocolEncoding as MEnc;
 
 use super::TerminalSelection;
 
-pub(super) fn ordered_selection(selection: TerminalSelection) -> ((u16, u16), (u16, u16)) {
+pub(super) fn ordered_selection(selection: TerminalSelection) -> ((i64, u16), (i64, u16)) {
     if selection.anchor <= selection.focus {
         (selection.anchor, selection.focus)
     } else {
